requirements.txt
flask>=3.0.0
sendgrid>=6.11.0
requests>=2.31.0
jinja2>=3.1.3
gunicorn>=21.2.0
APScheduler>=3.10.0
sentry-sdk[flask]>=1.40.0

# Dev/test dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
